Clean up template
