--- components/LoadingPage/LoadingPage.tsx
@@ -1,20 +1,0 @@
-import { CircularProgress, Typography } from "@material-ui/core";
-import React from "react";
-
-import { useStyles } from "./styles";
-
-function LoadingPage() {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.loaderContainer}>
-      <CircularProgress size={100} />
-
-      <Typography variant="subtitle1" className={classes.message}>
-        Attempting connection to Saleor Dashboard
-      </Typography>
-    </div>
-  );
-}
-
-export default LoadingPage;
