--- components/LoadingPage/styles.ts
@@ -1,14 +1,0 @@
-import { makeStyles } from "@saleor/macaw-ui";
-
-const useStyles = makeStyles((theme) => ({
-  loaderContainer: {
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-  },
-  message: {
-    marginTop: theme.spacing(4),
-  },
-}));
-
-export { useStyles };
