--- graphql/fragments/AddressFragment.graphql
@@ -1,17 +1,0 @@
-fragment AddressFragment on Address {
-  __typename
-  id
-  firstName
-  lastName
-  companyName
-  streetAddress1
-  streetAddress2
-  city
-  cityArea
-  postalCode
-  country {
-    code
-  }
-  countryArea
-  phone
-}
