--- graphql/mutations/UpdateAppMetadata.graphql
@@ -1,10 +1,0 @@
-mutation UpdateAppMetadata($id: ID!, $input: [MetadataInput!]!) {
-  updatePrivateMetadata(id: $id, input: $input) {
-    item {
-      privateMetadata {
-        key
-        value
-      }
-    }
-  }
-}
