--- graphql/queries/FetchVariousNumberOfOrders.graphql
@@ -1,26 +1,0 @@
-query FetchVariousNumberOfOrders($number_of_orders: Int!) {
-  orders(first: $number_of_orders) {
-    edges {
-      node {
-        id
-        number
-        created
-        userEmail
-        billingAddress {
-          ...AddressFragment
-        }
-        paymentStatus
-        status
-        total {
-          gross {
-            currency
-            amount
-          }
-        }
-        user {
-          id
-        }
-      }
-    }
-  }
-}
