--- graphql/subscriptions/OrderCreatedSubscription.graphql
@@ -1,10 +1,0 @@
-subscription OrderCreatedSubscription {
-  event {
-    ... on OrderCreated {
-      order {
-        id
-        userEmail
-      }
-    }
-  }
-}
