+++ src/lib/graphql.ts
