+++ src/pages/api/manifest.ts
@@ -1,0 +1,37 @@
+import { AppManifest } from "@saleor/app-sdk";
+import { createManifestHandler } from "@saleor/app-sdk/handlers/next";
+
+import packageJson from "../../../package.json";
+
+export default createManifestHandler({
+  async manifestFactory(context) {
+    const manifest: AppManifest = {
+      name: packageJson.name,
+      tokenTargetUrl: `${context.appBaseUrl}/api/register`,
+      appUrl: context.appBaseUrl,
+      permissions: [
+        /**
+         * Set permissions for app if needed
+         * https://docs.saleor.io/docs/3.x/developer/permissions
+         */
+      ],
+      id: "saleor.app",
+      version: packageJson.version,
+      webhooks: [
+        /**
+         * Configure webhooks here. They will be created in Saleor during installation
+         * Read more
+         * https://docs.saleor.io/docs/3.x/developer/api-reference/objects/webhook
+         */
+      ],
+      extensions: [
+        /**
+         * Optionally, extend Dashboard with custom UIs
+         * https://docs.saleor.io/docs/3.x/developer/extending/apps/extending-dashboard-with-apps
+         */
+      ],
+    };
+
+    return manifest;
+  },
+});
