+++ src/styles/globals.css
@@ -20,7 +20,3 @@ code::before {
   content: "$ ";
   opacity: 0.6;
 }
-
-li {
-  padding-bottom: 1rem;
-}
